Bug fix for connections to remote servers.
TCP servers may send messages in chunks, but the current implementation
assumes that a response is sent in one single message. This is usually
fine for localhost connections (where responses are small enough to be
sent to local sockets), but connections to hosts on the same network or
across the Internet may chunk the response into multiple messages. This
change uses `#read` instead of `#recv` which will block and buffer until
all bytes are received.